// src/prisma/client/models/ClientNotice.ts
export type ClientNoticeWhereInput = {
  AND?: Prisma.ClientNoticeWhereInput | Prisma.ClientNoticeWhereInput[]
  OR?: Prisma.ClientNoticeWhereInput[]
  NOT?: Prisma.ClientNoticeWhereInput | Prisma.ClientNoticeWhereInput[]
  id?: Prisma.IntFilter<"ClientNotice"> | number
  title?: Prisma.StringFilter<"ClientNotice"> | string
  content?: Prisma.StringFilter<"ClientNotice"> | string
  noticeType?: Prisma.IntFilter<"ClientNotice"> | number
  priorityLevel?: Prisma.IntFilter<"ClientNotice"> | number
  publishStartTime?: Prisma.DateTimeNullableFilter<"ClientNotice"> | Date | string | null
  publishEndTime?: Prisma.DateTimeNullableFilter<"ClientNotice"> | Date | string | null
  pageCode?: Prisma.StringNullableFilter<"ClientNotice"> | string | null
  popupBackgroundImage?: Prisma.StringNullableFilter<"ClientNotice"> | string | null
  isPublished?: Prisma.BoolFilter<"ClientNotice"> | boolean
  enablePlatform?: Prisma.IntFilter<"ClientNotice"> | number
  isPinned?: Prisma.BoolFilter<"ClientNotice"> | boolean
  showAsPopup?: Prisma.BoolFilter<"ClientNotice"> | boolean
  order?: Prisma.IntFilter<"ClientNotice"> | number
  readCount?: Prisma.IntFilter<"ClientNotice"> | number
  createdAt?: Prisma.DateTimeFilter<"ClientNotice"> | Date | string
  updatedAt?: Prisma.DateTimeFilter<"ClientNotice"> | Date | string
  deletedAt?: Prisma.DateTimeNullableFilter<"ClientNotice"> | Date | string | null
  clientPage?: Prisma.XOR<Prisma.ClientPageConfigNullableScalarRelationFilter, Prisma.ClientPageConfigWhereInput> | null
}